pyproject.toml
[tool.poetry]
name = "licenta"
version = "0.1.0"
description = ""
authors = ["aliaflore <100151385+aliaflore@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pdfplumber = "^0.11.0"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
numpy = "^1.26.4"
matplotlib = "^3.8.4"
opencv-python = "^4.9.0.80"
python-dotenv = "^1.0.1"
openai = "^1.23.6"
pypdf2 = "<3.0"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
celery = "^5.4.0"
redis = "^5.0.4"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
python-decouple = "^3.8"
eventlet = "^0.36.1"
gevent = "^24.2.1"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
watchdog = "^4.0.1"
requests = "^2.32.3"
camelot-py = "^0.11.0"
deep-translator = "^1.11.4"
thefuzz = "^0.22.1"
pymupdf = "^1.24.11"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
